+++ docs/api/conf.py
@@ -14,8 +14,8 @@ copyright = '2026, Interference Calculator Team'
 author = 'Interference Calculator Team'
 
 # The full version, including alpha/beta/rc tags
-release = '2.6.0'
-version = '2.6'
+release = '2.8.0'
+version = '2.8'
 
 # -- General configuration ---------------------------------------------------